pdp(chore): extract upload and create handlers into separate files
